refactor(edda-ledger): extract domain types to domain.rs (GH-378)
Move 9 domain-shaped types (PatternType, DetectedPattern, VillageStats, etc.)
from sqlite_store/types.rs to a new domain.rs module. Re-export from lib.rs
and update edda-serve to use the new import paths.

Phase 1 of 3: no behavioral change, only type relocation.
