+++ crates/edda-ledger/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod blob_meta;
 pub mod blob_store;
 pub mod device_token;
+pub mod domain;
 pub mod ledger;
 pub mod lock;
 pub mod paths;
@@ -15,12 +16,16 @@ pub use blob_store::{
     blob_put_classified, blob_put_if_large, blob_remove, blob_size, BlobInfo,
     SNAPSHOT_BLOB_THRESHOLD,
 };
+pub use domain::{
+    DayCount, DetectedPattern, DomainCount, ExecutionLinked, OutcomeMetrics,
+    PatternDetectionResult, PatternType, VillageStats, VillageStatsPeriod,
+};
 pub use ledger::Ledger;
 pub use lock::WorkspaceLock;
 pub use paths::EddaPaths;
 pub use sqlite_store::{
-    BundleRow, ChainEntry, DecideSnapshotRow, DecisionRow, DepRow, DetectedPattern, DeviceTokenRow,
-    ImportParams, PatternDetectionResult, PatternType, SuggestionRow, TaskBriefRow,
+    BundleRow, ChainEntry, DecideSnapshotRow, DecisionRow, DepRow, DeviceTokenRow, ImportParams,
+    SuggestionRow, TaskBriefRow,
 };
 pub use tombstone::{append_tombstone, list_tombstones, make_tombstone, DeleteReason, Tombstone};
 pub use view::DecisionView;
